+++ src/AzzyBot.Bot/Utilities/EmbedBuilder.cs
@@ -297,7 +297,7 @@ public static class EmbedBuilder
         return CreateBasicEmbed(title, description, DiscordColor.SpringGreen, new(file.Art), fields: fields);
     }
 
-    public static async Task<DiscordEmbed> BuildAzzyHardwareStatsEmbedAsync(Uri avaUrl)
+    public static async Task<DiscordEmbed> BuildAzzyHardwareStatsEmbedAsync(Uri avaUrl, int ping)
     {
         const string title = "AzzyBot Hardware Stats";
         const string notLinux = "To display more information you need to have a linux os.";
@@ -311,8 +311,14 @@ public static class EmbedBuilder
             ["Operating System"] = new(os, true),
             ["Architecture"] = new(osArch, true),
             ["Dockerized?"] = new(Misc.GetReadableBool(isDocker, ReadableBool.YesNo), true),
-            ["System Uptime"] = new($"<t:{uptime}>")
+            ["System Uptime"] = new($"<t:{uptime}>", true)
         };
+
+        if (ping is not 0)
+        {
+            fields.Add("Ping", new($"{ping} ms", true));
+            fields.Add("\u200b", new("\u200b", true)); // Empty field to make the embed look better
+        }
 
         if (!HardwareStats.CheckIfLinuxOs)
             return CreateBasicEmbed(title, color: DiscordColor.Orange, footerText: notLinux, fields: fields);