+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Improvements
 - Greatly refactored the handling of DSharpPlus related stuff
 - The usual connection error message of updating AzuraCast should not be shown anymore
+- `core stats hardware` now shows the latency to discord
 
 ### Fixes
 - MusicStreaming commands exit correctly now when there is an error